Avoid accessing OCaml heap in the blocking section
We were traversing the sigprocmask argument too late, after entering the
blocking section. Now we traverse the argument earlier, when considering the
other arguments, and only do the actual sigmask operations in the blocking
section.

This costs an extra system call (to pthread_sigmask) in the non-posix_spawn
case, because previously the code was using a single pthread_sigmask call to
both get the current sigmask and block all signals. Now we use it twice: before
the blocking section to get the current mask, and then inside the blocking
section to block all signals.

Signed-off-by: Jesse Tov <jtov@janestreet.com>